training/requirements.txt
torch
pillow
numpy
tqdm
onnx
